requirements.txt
aiohappyeyeballs
aiohttp
aioice
aiortc
aiosignal
altair
annotated-types
anyio
appdirs
asgiref
asttokens
attrs
auth0-python
av
backoff
bcrypt
beautifulsoup4
blinker
bs4
build
cachetools
certifi
cffi
chardet
charset-normalizer
chroma-hnswlib
chromadb
click
colorama
coloredlogs
comm
contourpy
crewai
cryptography
cycler
dataclasses-json
debugpy
decorator
Deprecated
distro
dnspython
docstring_parser
durationpy
email_validator
entrypoints
et_xmlfile
executing
faiss-cpu
Faker
fastapi
favicon
filelock
filetype
firecrawl
flatbuffers
fonttools
frozenlist
fsspec
gitdb
GitPython
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-crc32c
google-generativeai
google_search_results
googleapis-common-protos
greenlet
groq
grpcio
grpcio-status
h11
htbuilder
httpcore
httplib2
httptools
httpx
httpx-sse
huggingface-hub
humanfriendly
idna
ifaddr
importlib_metadata
importlib_resources
instructor
ipykernel
ipython
ipython_pygments_lexers
jedi
Jinja2
jiter
joblib
json5
json_repair
jsonpatch
jsonpickle
jsonpointer
jsonref
jsonschema
jsonschema-specifications
jupyter_client
jupyter_core
kiwisolver
kubernetes
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-groq
langchain-text-splitters
langsmith
litellm
lxml
Markdown
markdown-it-py
markdownlit
MarkupSafe
marshmallow
matplotlib
matplotlib-inline
mdurl
mmh3
monotonic
motor
mpmath
multidict
mypy_extensions
narwhals
nest-asyncio
networkx
numpy
oauthlib
onnxruntime
openai
openpyxl
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-fastapi
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-util-http
orjson
outcome
overrides
packaging
pandas
parso
pdfminer.six
pdfplumber
pillow
platformdirs
playwright
plotly
posthog
prometheus_client
prompt_toolkit
propcache
proto-plus
protobuf
psutil
pure_eval
pyarrow
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic-settings
pydantic_core
pydeck
pyee
Pygments
PyJWT
pylibsrtp
pymdown-extensions
pymongo
PyMuPDF
pyOpenSSL
pyparsing
PyPDF2
pypdfium2
PyPika
pyproject_hooks
pyreadline3
PySocks
python-dateutil
python-docx
python-dotenv
pytz
ppyvis
pywin32
PyYAML
pyzmq
referencing
regex
reportlab
requests
requests-oauthlib
requests-toolbelt
rich
rpds-py
rsa
safetensors
scikit-learn
scipy
selenium
sentence-transformers
setuptools
shellingham
six
smmap
sniffio
sortedcontainers
soupsieve
SQLAlchemy
st-annotated-text
st-theme
stack-data
starlette
streamlit
streamlit-audiorec
streamlit-avatar
streamlit-camera-input-live
streamlit-card
streamlit-embedcode
streamlit-extras
streamlit-faker
streamlit-image-coordinates
streamlit-keyup
streamlit-lottie
streamlit-toggle-switch
streamlit-vertical-slider
streamlit-webrtc
sympy
tenacity
threadpoolctl
tiktoken
tokenizers
toml
tomli
tomli_w
torch
tornado
tqdm
traitlets
transformers
trio
trio-websocket
typer
typing-inspect
typing-inspection
typing_extensions
tzdata
uritemplate
urllib3
uv
uvicorn
validators
watchdog
watchfiles
wcwidth
webdriver-manager
websocket-client
websockets
wrapt
wsproto
yarl
zipp
zstandard
pysqlite3-binary
